Create Dotilist.html
I’m excited to share my latest project, a to-do list app developed using HTML, JavaScript, and CSS! This app features a basic to-do list with an easy-to-use interface, an add button to add tasks, a remove button to delete tasks, and an edit button to change tasks. Building this app deepened my skills in HTML, JavaScript, and CSS, and also introduced me to Android development.